MP-060 D.5: split VM dispatch helpers and enforce LOC budget
