+++ crates/trust-runtime/src/runtime/vm/dispatch.rs
@@ -2,20 +2,25 @@ use std::time::Instant;
 
 use smol_str::SmolStr;
 
-use crate::bytecode::{TypeData, TypeTable};
 use crate::debug::DebugHook;
 use crate::error::RuntimeError;
-use crate::eval::ops::{apply_binary, apply_unary, BinaryOp, UnaryOp};
-use crate::memory::{FrameId, InstanceId, MemoryLocation};
+use crate::eval::ops::{BinaryOp, UnaryOp};
+use crate::memory::InstanceId;
 use crate::task::ProgramDef;
-use crate::value::{size_of_value, RefSegment, SizeOfError, Value, ValueRef};
+use crate::value::{size_of_value, Value, ValueRef};
 
 use super::super::core::Runtime;
-use super::call::{execute_native_call, push_call_frame, VM_LOCAL_SENTINEL_FRAME_ID};
+use super::call::{execute_native_call, push_call_frame};
+use super::dispatch_ops::{apply_jump, execute_binary, execute_unary, read_i32, read_u32};
+use super::dispatch_refs::{
+    dynamic_load_ref, dynamic_ref_field, dynamic_ref_index, dynamic_store_ref, index_to_i64,
+    load_ref, load_ref_addr, pop_reference, store_ref,
+};
+use super::dispatch_sizeof::{sizeof_error_to_runtime, sizeof_type_from_table};
 use super::errors::VmTrap;
-use super::frames::{FrameStack, VmFrame};
+use super::frames::FrameStack;
 use super::stack::OperandStack;
-use super::{VmModule, VmRef};
+use super::VmModule;
 
 pub(super) fn execute_program(
     runtime: &mut Runtime,
@@ -412,521 +417,4 @@ fn vm_statement_location(
 ) -> Option<crate::debug::SourceLocation> {
     let source = module.debug_map.source_by_pc.get(&(pou_id, pc as u32))?;
     runtime.resolve_vm_debug_location(source.file.as_str(), source.line, source.column)
-}
-
-fn execute_unary(stack: &mut OperandStack, op: UnaryOp) -> Result<(), VmTrap> {
-    let value = stack.pop()?;
-    let result = apply_unary(op, value)?;
-    stack.push(result)
-}
-
-fn execute_binary(runtime: &Runtime, stack: &mut OperandStack, op: BinaryOp) -> Result<(), VmTrap> {
-    let (left, right) = stack.pop_pair()?;
-    let result = apply_binary(op, left, right, &runtime.profile)?;
-    stack.push(result)
-}
-
-fn load_ref(
-    runtime: &Runtime,
-    module: &VmModule,
-    frames: &FrameStack,
-    ref_idx: u32,
-) -> Result<Value, VmTrap> {
-    let reference = module
-        .refs
-        .get(ref_idx as usize)
-        .ok_or(VmTrap::InvalidRefIndex(ref_idx))?;
-
-    match reference {
-        VmRef::Local { path, .. } => {
-            if !path.is_empty() {
-                return Err(VmTrap::UnsupportedRefLocation("local-path"));
-            }
-            let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-            frame.load_local(ref_idx)
-        }
-        _ => {
-            let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-            let (location, offset, path) = runtime_access_target(reference, frame)?;
-            runtime
-                .storage
-                .read_by_ref_parts(location, offset, path)
-                .cloned()
-                .ok_or(VmTrap::NullReference)
-        }
-    }
-}
-
-fn load_ref_addr(module: &VmModule, frames: &FrameStack, ref_idx: u32) -> Result<ValueRef, VmTrap> {
-    let reference = module
-        .refs
-        .get(ref_idx as usize)
-        .ok_or(VmTrap::InvalidRefIndex(ref_idx))?;
-    match reference {
-        VmRef::Local { path, .. } => {
-            if !path.is_empty() {
-                return Err(VmTrap::UnsupportedRefLocation("local-path"));
-            }
-            let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-            let local_slot = frame.local_slot_index(ref_idx)?;
-            Ok(ValueRef {
-                location: MemoryLocation::Local(FrameId(VM_LOCAL_SENTINEL_FRAME_ID)),
-                offset: local_slot,
-                path: Vec::new(),
-            })
-        }
-        _ => {
-            let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-            let (location, offset, path) = runtime_access_target(reference, frame)?;
-            Ok(ValueRef {
-                location,
-                offset,
-                path: path.to_vec(),
-            })
-        }
-    }
-}
-
-fn store_ref(
-    runtime: &mut Runtime,
-    module: &VmModule,
-    frames: &mut FrameStack,
-    ref_idx: u32,
-    value: Value,
-) -> Result<(), VmTrap> {
-    let reference = module
-        .refs
-        .get(ref_idx as usize)
-        .ok_or(VmTrap::InvalidRefIndex(ref_idx))?;
-
-    match reference {
-        VmRef::Local { path, .. } => {
-            if !path.is_empty() {
-                return Err(VmTrap::UnsupportedRefLocation("local-path"));
-            }
-            let frame = frames.current_mut().ok_or(VmTrap::CallStackUnderflow)?;
-            frame.store_local(ref_idx, value)
-        }
-        _ => {
-            let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-            let (location, offset, path) = runtime_access_target(reference, frame)?;
-            if runtime
-                .storage
-                .write_by_ref_parts(location, offset, path, value)
-            {
-                Ok(())
-            } else {
-                Err(VmTrap::NullReference)
-            }
-        }
-    }
-}
-
-fn pop_reference(stack: &mut OperandStack) -> Result<ValueRef, VmTrap> {
-    let value = stack.pop()?;
-    match value {
-        Value::Reference(Some(reference)) => Ok(reference),
-        Value::Reference(None) => Err(VmTrap::NullReference),
-        _ => Err(VmTrap::Runtime(RuntimeError::TypeMismatch)),
-    }
-}
-
-fn dynamic_ref_field(
-    runtime: &Runtime,
-    frames: &FrameStack,
-    mut reference: ValueRef,
-    field: SmolStr,
-) -> Result<ValueRef, VmTrap> {
-    let target = dynamic_load_ref(runtime, frames, &reference)?;
-    match target {
-        Value::Struct(struct_value) => {
-            if !struct_value.fields.contains_key(field.as_str()) {
-                return Err(VmTrap::Runtime(RuntimeError::UndefinedField(field)));
-            }
-            reference.path.push(RefSegment::Field(field));
-            Ok(reference)
-        }
-        Value::Instance(instance_id) => runtime
-            .storage
-            .ref_for_instance_recursive(instance_id, field.as_str())
-            .ok_or(VmTrap::Runtime(RuntimeError::UndefinedField(field))),
-        _ => Err(VmTrap::Runtime(RuntimeError::TypeMismatch)),
-    }
-}
-
-fn dynamic_ref_index(
-    runtime: &Runtime,
-    frames: &FrameStack,
-    mut reference: ValueRef,
-    index: i64,
-) -> Result<ValueRef, VmTrap> {
-    let target = dynamic_load_ref(runtime, frames, &reference)?;
-    match target {
-        Value::Array(array) => {
-            if array.dimensions.len() != 1 {
-                return Err(VmTrap::Runtime(RuntimeError::TypeMismatch));
-            }
-            let (lower, upper) = array.dimensions[0];
-            if index < lower || index > upper {
-                return Err(VmTrap::Runtime(RuntimeError::IndexOutOfBounds {
-                    index,
-                    lower,
-                    upper,
-                }));
-            }
-            reference.path.push(RefSegment::Index(vec![index]));
-            Ok(reference)
-        }
-        _ => Err(VmTrap::Runtime(RuntimeError::TypeMismatch)),
-    }
-}
-
-fn dynamic_load_ref(
-    runtime: &Runtime,
-    frames: &FrameStack,
-    reference: &ValueRef,
-) -> Result<Value, VmTrap> {
-    if matches!(
-        reference.location,
-        MemoryLocation::Local(FrameId(VM_LOCAL_SENTINEL_FRAME_ID))
-    ) {
-        let frame = frames.current().ok_or(VmTrap::CallStackUnderflow)?;
-        return read_vm_local_ref(frame, reference.offset, &reference.path);
-    }
-    runtime
-        .storage
-        .read_by_ref(reference.clone())
-        .cloned()
-        .ok_or(VmTrap::NullReference)
-}
-
-fn dynamic_store_ref(
-    runtime: &mut Runtime,
-    frames: &mut FrameStack,
-    reference: &ValueRef,
-    value: Value,
-) -> Result<(), VmTrap> {
-    if matches!(
-        reference.location,
-        MemoryLocation::Local(FrameId(VM_LOCAL_SENTINEL_FRAME_ID))
-    ) {
-        let frame = frames.current_mut().ok_or(VmTrap::CallStackUnderflow)?;
-        return write_vm_local_ref(frame, reference.offset, &reference.path, value);
-    }
-    if runtime.storage.write_by_ref(reference.clone(), value) {
-        Ok(())
-    } else {
-        Err(VmTrap::NullReference)
-    }
-}
-
-fn read_vm_local_ref(frame: &VmFrame, offset: usize, path: &[RefSegment]) -> Result<Value, VmTrap> {
-    let root = frame.locals.get(offset).ok_or(VmTrap::NullReference)?;
-    read_value_path(root, path)
-        .cloned()
-        .ok_or(VmTrap::NullReference)
-}
-
-fn write_vm_local_ref(
-    frame: &mut VmFrame,
-    offset: usize,
-    path: &[RefSegment],
-    value: Value,
-) -> Result<(), VmTrap> {
-    let root = frame.locals.get_mut(offset).ok_or(VmTrap::NullReference)?;
-    if write_value_path(root, path, value) {
-        Ok(())
-    } else {
-        Err(VmTrap::NullReference)
-    }
-}
-
-fn read_value_path<'a>(value: &'a Value, path: &[RefSegment]) -> Option<&'a Value> {
-    if path.is_empty() {
-        return Some(value);
-    }
-    match &path[0] {
-        RefSegment::Field(name) => match value {
-            Value::Struct(struct_value) => struct_value
-                .fields
-                .get(name)
-                .and_then(|field| read_value_path(field, &path[1..])),
-            _ => None,
-        },
-        RefSegment::Index(indices) => match value {
-            Value::Array(array) => {
-                let offset = array_offset_i64(&array.dimensions, indices)?;
-                array
-                    .elements
-                    .get(offset)
-                    .and_then(|element| read_value_path(element, &path[1..]))
-            }
-            _ => None,
-        },
-    }
-}
-
-fn write_value_path(target: &mut Value, path: &[RefSegment], value: Value) -> bool {
-    if path.is_empty() {
-        *target = value;
-        return true;
-    }
-
-    match &path[0] {
-        RefSegment::Field(name) => match target {
-            Value::Struct(struct_value) => struct_value
-                .fields
-                .get_mut(name)
-                .map(|field| write_value_path(field, &path[1..], value))
-                .unwrap_or(false),
-            _ => false,
-        },
-        RefSegment::Index(indices) => match target {
-            Value::Array(array) => {
-                let offset = match array_offset_i64(&array.dimensions, indices) {
-                    Some(offset) => offset,
-                    None => return false,
-                };
-                array
-                    .elements
-                    .get_mut(offset)
-                    .map(|element| write_value_path(element, &path[1..], value))
-                    .unwrap_or(false)
-            }
-            _ => false,
-        },
-    }
-}
-
-fn array_offset_i64(dimensions: &[(i64, i64)], indices: &[i64]) -> Option<usize> {
-    if dimensions.len() != indices.len() {
-        return None;
-    }
-    let mut offset: i128 = 0;
-    let mut stride: i128 = 1;
-    for ((lower, upper), index) in dimensions.iter().zip(indices).rev() {
-        if index < lower || index > upper {
-            return None;
-        }
-        let len = (*upper - *lower + 1) as i128;
-        offset += (index - *lower) as i128 * stride;
-        stride *= len;
-    }
-    usize::try_from(offset).ok()
-}
-
-fn index_to_i64(value: Value) -> Result<i64, VmTrap> {
-    match value {
-        Value::SInt(v) => Ok(v as i64),
-        Value::Int(v) => Ok(v as i64),
-        Value::DInt(v) => Ok(v as i64),
-        Value::LInt(v) => Ok(v),
-        Value::USInt(v) => Ok(v as i64),
-        Value::UInt(v) => Ok(v as i64),
-        Value::UDInt(v) => Ok(v as i64),
-        Value::ULInt(v) => {
-            i64::try_from(v).map_err(|_| VmTrap::Runtime(RuntimeError::TypeMismatch))
-        }
-        _ => Err(VmTrap::Runtime(RuntimeError::TypeMismatch)),
-    }
-}
-
-fn sizeof_error_to_runtime(err: SizeOfError) -> RuntimeError {
-    match err {
-        SizeOfError::Overflow => RuntimeError::Overflow,
-        SizeOfError::UnknownType | SizeOfError::UnsupportedType => RuntimeError::TypeMismatch,
-    }
-}
-
-fn sizeof_type_from_table(types: &TypeTable, type_idx: u32) -> Result<u64, RuntimeError> {
-    let mut stack = Vec::new();
-    sizeof_type_from_table_inner(types, type_idx, &mut stack)
-}
-
-const SIZEOF_TYPE_MAX_DEPTH: usize = 128;
-
-fn sizeof_type_from_table_inner(
-    types: &TypeTable,
-    type_idx: u32,
-    stack: &mut Vec<u32>,
-) -> Result<u64, RuntimeError> {
-    if stack.len() >= SIZEOF_TYPE_MAX_DEPTH {
-        return Err(RuntimeError::InvalidBytecode(
-            format!(
-                "SIZEOF type nesting exceeds max depth {SIZEOF_TYPE_MAX_DEPTH} at index {type_idx}"
-            )
-            .into(),
-        ));
-    }
-    if stack.contains(&type_idx) {
-        return Err(RuntimeError::InvalidBytecode(
-            format!("SIZEOF type recursion at index {type_idx}").into(),
-        ));
-    }
-    let entry = types.entries.get(type_idx as usize).ok_or_else(|| {
-        RuntimeError::InvalidBytecode(format!("invalid type index {type_idx} for SIZEOF").into())
-    })?;
-    stack.push(type_idx);
-    let result = (|| match &entry.data {
-        TypeData::Primitive {
-            prim_id,
-            max_length,
-        } => sizeof_primitive_type(*prim_id, *max_length),
-        TypeData::Array { elem_type_id, dims } => {
-            let elem_size = sizeof_type_from_table_inner(types, *elem_type_id, stack)?;
-            let len = type_array_len(dims).ok_or(RuntimeError::TypeMismatch)?;
-            elem_size.checked_mul(len).ok_or(RuntimeError::Overflow)
-        }
-        TypeData::Struct { fields } => {
-            let mut total = 0u64;
-            for field in fields {
-                let size = sizeof_type_from_table_inner(types, field.type_id, stack)?;
-                total = total.checked_add(size).ok_or(RuntimeError::Overflow)?;
-            }
-            Ok(total)
-        }
-        TypeData::Union { fields } => {
-            let mut max_size = 0u64;
-            for field in fields {
-                let size = sizeof_type_from_table_inner(types, field.type_id, stack)?;
-                max_size = max_size.max(size);
-            }
-            Ok(max_size)
-        }
-        TypeData::Enum { base_type_id, .. } => {
-            sizeof_type_from_table_inner(types, *base_type_id, stack)
-        }
-        TypeData::Alias { target_type_id } => {
-            sizeof_type_from_table_inner(types, *target_type_id, stack)
-        }
-        TypeData::Subrange { base_type_id, .. } => {
-            sizeof_type_from_table_inner(types, *base_type_id, stack)
-        }
-        TypeData::Reference { .. } => {
-            u64::try_from(std::mem::size_of::<ValueRef>()).map_err(|_| RuntimeError::Overflow)
-        }
-        TypeData::Pou { .. } | TypeData::Interface { .. } => Err(RuntimeError::TypeMismatch),
-    })();
-    let _ = stack.pop();
-    result
-}
-
-fn sizeof_primitive_type(prim_id: u16, max_length: u16) -> Result<u64, RuntimeError> {
-    match prim_id {
-        1 | 2 | 6 | 10 | 26 => Ok(1),
-        3 | 7 | 11 | 27 => Ok(2),
-        4 | 8 | 12 | 14 => Ok(4),
-        5 | 9 | 13 | 15 => Ok(8),
-        16 | 18 | 20 | 22 => Ok(4),
-        17 | 19 | 21 | 23 => Ok(8),
-        24 => {
-            if max_length == 0 {
-                Err(RuntimeError::TypeMismatch)
-            } else {
-                Ok(u64::from(max_length))
-            }
-        }
-        25 => {
-            if max_length == 0 {
-                Err(RuntimeError::TypeMismatch)
-            } else {
-                u64::from(max_length)
-                    .checked_mul(2)
-                    .ok_or(RuntimeError::Overflow)
-            }
-        }
-        other => Err(RuntimeError::InvalidBytecode(
-            format!("unsupported primitive type id {other} for SIZEOF").into(),
-        )),
-    }
-}
-
-fn type_array_len(dimensions: &[(i64, i64)]) -> Option<u64> {
-    let mut total: i128 = 1;
-    for (lower, upper) in dimensions {
-        let len = i128::from(*upper) - i128::from(*lower) + 1;
-        if len <= 0 {
-            return None;
-        }
-        total = total.checked_mul(len)?;
-    }
-    u64::try_from(total).ok()
-}
-
-fn runtime_access_target<'a>(
-    reference: &'a VmRef,
-    frame: &VmFrame,
-) -> Result<(MemoryLocation, usize, &'a [crate::value::RefSegment]), VmTrap> {
-    match reference {
-        VmRef::Global { offset, path } => Ok((MemoryLocation::Global, *offset, path.as_slice())),
-        VmRef::Instance {
-            owner_instance_id,
-            offset,
-            path,
-        } => {
-            let runtime_owner = if frame.instance_owner == Some(*owner_instance_id) {
-                frame
-                    .runtime_instance
-                    .unwrap_or(InstanceId(*owner_instance_id))
-            } else {
-                InstanceId(*owner_instance_id)
-            };
-            Ok((
-                MemoryLocation::Instance(runtime_owner),
-                *offset,
-                path.as_slice(),
-            ))
-        }
-        VmRef::Local {
-            owner_frame_id,
-            offset,
-            path,
-        } => Err(VmTrap::UnsupportedRefLocation(if path.is_empty() {
-            let _ = owner_frame_id;
-            let _ = offset;
-            "local"
-        } else {
-            "local-path"
-        })),
-        VmRef::Retain { offset, path } => {
-            let _ = (offset, path);
-            Err(VmTrap::UnsupportedRefLocation("retain"))
-        }
-        VmRef::Io { area, offset, path } => {
-            let _ = (area, offset, path);
-            Err(VmTrap::UnsupportedRefLocation("io"))
-        }
-    }
-}
-
-fn apply_jump(pc: &mut usize, offset: i32, frame: &VmFrame) -> Result<(), VmTrap> {
-    let base = *pc as i64;
-    let target = base + i64::from(offset);
-    if target < frame.code_start as i64 || target > frame.code_end as i64 {
-        return Err(VmTrap::InvalidJumpTarget(target));
-    }
-    *pc = target as usize;
-    Ok(())
-}
-
-fn read_u32(code: &[u8], pc: &mut usize) -> Result<u32, VmTrap> {
-    if *pc + 4 > code.len() {
-        return Err(VmTrap::BytecodeDecode(
-            "vm operand read overflow (u32)".into(),
-        ));
-    }
-    let bytes = [code[*pc], code[*pc + 1], code[*pc + 2], code[*pc + 3]];
-    *pc += 4;
-    Ok(u32::from_le_bytes(bytes))
-}
-
-fn read_i32(code: &[u8], pc: &mut usize) -> Result<i32, VmTrap> {
-    if *pc + 4 > code.len() {
-        return Err(VmTrap::BytecodeDecode(
-            "vm operand read overflow (i32)".into(),
-        ));
-    }
-    let bytes = [code[*pc], code[*pc + 1], code[*pc + 2], code[*pc + 3]];
-    *pc += 4;
-    Ok(i32::from_le_bytes(bytes))
 }
